Replace nameOverride by fullnameOverride for  kube-prometheus-stack dependency
kube-prometheus-stack chart concatenates <.Release.Name> with <nameOverride>
when <nameOverride> is used. Then the <.Release.Name>-<nameOverride> k8s
name is truncates to 26 characters. Therefore we can not predict the
service name of kube-prometheus-stack objects when long <.Release.Name>
is used. It causes that we can not define additionalDataSources urls for
Grafana.

Signed-off-by: Matej Feder <matej.feder@ifne.eu>